Move Context Menu class in core folder
