+++ lib/testing/src/lib/core/pages/public-api.ts
@@ -32,3 +32,4 @@ export * from './config-editor-page';
 export * from './about.page';
 export * from './snackbar.page';
 export * from './data-table/public-api';
+export * from './context-menu.page';
